+++ src/bfs/bfs_kernels.cu
@@ -207,7 +207,7 @@ __global__ void BFS64_Filter_Kernel(const uint32 *partitionList,
   uint32 partition = partitionList[0];
   // uint32 touch;
   //  Start Edge
-  uint32 startEdge = d_offsets[d_partitionsOffsets[partition]];
+  uint64 startEdge = d_offsets[d_partitionsOffsets[partition]];
 
   // End Edge
   // uint32 endEdge = d_offsets[d_partitionsOffsets[partition + 1]];
@@ -228,13 +228,13 @@ __global__ void BFS64_Filter_Kernel(const uint32 *partitionList,
 
     d_filterFrontier[warpIdx] = 0;
 
-    uint32 newValue = d_values[warpIdx] + 1;
+    uint64 newValue = d_values[warpIdx] + 1;
 
     const uint64 start = d_offsets[warpIdx] - startEdge;
     const uint64 end = d_offsets[warpIdx + 1] - startEdge;
 
     for (uint64 i = start + laneIdx; i < end; i += WARP_SIZE) {
-      uint32 neighborId = d_filterEdges[i];
+      uint64 neighborId = d_filterEdges[i];
 
       // If this new path has lower cost than the previous then change and add
       // the neighbor to the frontier
@@ -389,7 +389,7 @@ __global__ void BFS64_NeighborFilter_Kernel(const uint32 *partitionList,
   uint32 partition = partitionList[0];
   // uint32 touch;
   //  Start Edge
-  uint32 startEdge = d_offsets[d_partitionsOffsets[partition]];
+  uint64 startEdge = d_offsets[d_partitionsOffsets[partition]];
 
   // End Edge
   // uint32 endEdge = d_offsets[d_partitionsOffsets[partition + 1]];
@@ -409,13 +409,13 @@ __global__ void BFS64_NeighborFilter_Kernel(const uint32 *partitionList,
 
     d_filterFrontier[warpIdx] = 0;
 
-    uint32 newValue = d_values[warpIdx] + 1;
+    uint64 newValue = d_values[warpIdx] + 1;
 
     const uint64 start = d_offsets[warpIdx] - startEdge;
     const uint64 end = d_offsets[warpIdx + 1] - startEdge;
 
     for (uint64 i = start + laneIdx; i < end; i += WARP_SIZE) {
-      uint32 neighborId = d_filterEdges[i];
+      uint64 neighborId = d_filterEdges[i];
 
       // If this new path has lower cost than the previous then change and add
       // the neighbor to the frontier
